refactor(Coupon) : 쿠폰 id 하드코딩 -> enum으로 임시 대체
