+++ backend-server/src/main/java/shop/bluebooktle/backend/coupon/dto/CouponSearchRequest.java
@@ -2,7 +2,7 @@ package shop.bluebooktle.backend.coupon.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import shop.bluebooktle.common.domain.CouponTypeTarget;
+import shop.bluebooktle.common.domain.coupon.CouponTypeTarget;
 
 @Getter
 @Setter
