+++ backend-server/src/main/java/shop/bluebooktle/backend/coupon/entity/CouponType.java
@@ -18,7 +18,7 @@ import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shop.bluebooktle.common.domain.CouponTypeTarget;
+import shop.bluebooktle.common.domain.coupon.CouponTypeTarget;
 import shop.bluebooktle.common.entity.BaseEntity;
 
 @Entity
