+++ backend-server/src/main/java/shop/bluebooktle/backend/coupon/service/impl/CouponServiceImpl.java
@@ -20,7 +20,7 @@ import shop.bluebooktle.backend.coupon.repository.CategoryCouponRepository;
 import shop.bluebooktle.backend.coupon.repository.CouponRepository;
 import shop.bluebooktle.backend.coupon.repository.CouponTypeRepository;
 import shop.bluebooktle.backend.coupon.service.CouponService;
-import shop.bluebooktle.common.domain.CouponTypeTarget;
+import shop.bluebooktle.common.domain.coupon.CouponTypeTarget;
 import shop.bluebooktle.common.dto.coupon.request.CouponRegisterRequest;
 import shop.bluebooktle.common.dto.coupon.response.CouponResponse;
 import shop.bluebooktle.common.exception.book.BookNotFoundException;
