--- common-module/src/main/java/shop/bluebooktle/common/domain/CouponTypeTarget.java
@@ -1,5 +1,0 @@
-package shop.bluebooktle.common.domain;
-
-public enum CouponTypeTarget {
-	ORDER, BOOK
-}
