+++ common-module/src/main/java/shop/bluebooktle/common/dto/coupon/response/CouponResponse.java
@@ -7,7 +7,7 @@ import com.querydsl.core.annotations.QueryProjection;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shop.bluebooktle.common.domain.CouponTypeTarget;
+import shop.bluebooktle.common.domain.coupon.CouponTypeTarget;
 
 @Getter
 @NoArgsConstructor
